Move splitting log to logs module
