+++ core/core.py
@@ -1,12 +1,10 @@
-from sklearn.model_selection import train_test_split
-
 from core.classification import classifier
 from core.constants import NEXT_ACTIVITY, \
     CLASSIFICATION, REGRESSION
 from core.next_activity import next_activity
 from core.regression import regression
 from encoders.common import encode_logs
-from logs.file_service import get_logs
+from logs.splitting import prepare_logs
 
 
 def calculate(job):
@@ -33,25 +31,6 @@ def calculate(job):
         raise ValueError("Type not supported", job['type'])
     print("End job {}, {} . Results {}".format(job['type'], get_run(job), results))
     return results
-
-
-def prepare_logs(split: dict):
-    """Returns training_log and test_log"""
-    if split['type'] == 'single':
-        log = get_logs(split['original_log_path'])[0]
-        training_log, test_log = split_log(log)
-        print("Loaded single log from {}".format(split['original_log_path']))
-    else:
-        # Have to use sklearn to convert some internal data types
-        training_log, _ = train_test_split(get_logs(split['training_log_path'])[0], test_size=0)
-        test_log, _ = train_test_split(get_logs(split['test_log_path'])[0], test_size=0)
-        print("Loaded double logs from {} and {}.".format(split['training_log_path'], split['test_log_path']))
-    return training_log, test_log
-
-
-def split_log(log: list, test_size=0.2, random_state=4):
-    training_log, test_log = train_test_split(log, test_size=test_size, random_state=random_state)
-    return training_log, test_log
 
 
 def get_run(job):
